Fixes comments, colocates tests and resolver code with endpoints
