+++ src/sentry/web/frontend/base.py
@@ -155,8 +155,8 @@ def cell_silo_view(
     By default, if a request is received and the application is not in CELL
     mode 404s will be returned.
 
-    Optionally, a resolver class can be specified to allow a Control to dispatch
-    a request to the correct cell, if possible.
+    Optionally, a resolver class can be specified to allow Control Silo to
+    dispatch a request to the correct cell, if possible.
     """
     limiter = CellSiloView(internal=False, control_silo_resolver=control_silo_resolver)
     if decorated_obj is not None:
@@ -178,8 +178,8 @@ def internal_cell_silo_view(
     and are not accessible via cell routing.
     This is generally for debug/development views.
 
-    Optionally, a resolver class can be specified to allow a Control to dispatch
-    a request to the correct cell, if possible.
+    Optionally, a resolver class can be specified to allow Control Silo to
+    dispatch a request to the correct cell, if possible.
     """
     limiter = CellSiloView(internal=True, control_silo_resolver=control_silo_resolver)
     if decorated_obj is not None:
